move PaginationWrapper to sharedComponents directory
