+++ apps/src/levelbuilder/code-docs-editor/ProgrammingExpressionsTable.jsx
@@ -5,7 +5,7 @@ import * as Table from 'reactabular-table';
 
 import Button from '@cdo/apps/legacySharedComponents/Button';
 import StylizedBaseDialog from '@cdo/apps/sharedComponents/StylizedBaseDialog';
-import PaginationWrapper from '@cdo/apps/templates/PaginationWrapper';
+import PaginationWrapper from '@cdo/apps/sharedComponents/PaginationWrapper';
 
 import CloneProgrammingExpressionDialog from './CloneProgrammingExpressionDialog';
