+++ apps/src/sharedComponents/README.md
@@ -2,11 +2,10 @@ This folder contains UI components that are used in multiple places
 across the site.
 
 Components in this folder do not have Design System counterparts,
-but are still generic enough to be reusable. If there is a DSCO component that
+but are still generic enough to be reusable. If there is a DSCO or MUI component that
 can suit your needs, use it rather than a component from this directory.
 
 We're in the process of inventorying, consolidating and deleting components,
 so please be mindful of clean up opportunities when using these components.
 
-You can find more details about modernization efforts [here](https://docs.google.com/document/d/1Yxek5dqYKLJUraB17yCdEFLlJvlOiB1zhSWZAuu3zPU/edit#heading=h.18w2m61099gz)
-and log clean-up tasks [here](https://codedotorg.atlassian.net/browse/XTEAM-426).
+You can find more details about modernization efforts [here](https://docs.google.com/document/d/1Yxek5dqYKLJUraB17yCdEFLlJvlOiB1zhSWZAuu3zPU/edit#heading=h.18w2m61099gz).
